requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
pandas
matplotlib
